Split clear command
