+++ src/test/java/seedu/address/logic/commands/teacher/ClearTeacherCommandTest.java
@@ -1,0 +1,30 @@
+package seedu.address.logic.commands.teacher;
+
+import org.junit.jupiter.api.Test;
+import seedu.address.model.Model;
+import seedu.address.model.ModelManager;
+import seedu.address.model.UserPrefs;
+
+import static seedu.address.logic.commands.CommandTestUtil.assertCommandSuccess;
+import static seedu.address.testutil.TypicalPersons.getTypicalAddressBook;
+import static seedu.address.testutil.TypicalPersons.getTypicalAddressBookStudents;
+
+public class ClearTeacherCommandTest {
+
+    @Test
+    public void execute_emptyAddressBook_success() {
+        Model model = new ModelManager();
+        Model expectedModel = new ModelManager();
+
+        assertCommandSuccess(new ClearTeacherCommand(), model, ClearTeacherCommand.MESSAGE_SUCCESS, expectedModel);
+    }
+
+    @Test
+    public void execute_nonEmptyAddressBook_success() {
+        Model model = new ModelManager(getTypicalAddressBook(), new UserPrefs());
+        Model expectedModel = new ModelManager(getTypicalAddressBookStudents(), new UserPrefs());
+
+        // after clearing teachers, should only be left with students.
+        assertCommandSuccess(new ClearTeacherCommand(), model, ClearTeacherCommand.MESSAGE_SUCCESS, expectedModel);
+    }
+}
